requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
astroid==2.11.4
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.2
dill==0.3.4
h11==0.13.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
outcome==1.1.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pylint==2.13.8
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.2
python-dotenv==0.20.0
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wrapt==1.14.1
wsproto==1.1.0
